third_party/WebKit/LayoutTests/fast/css/style-scoped/style-scoped-shadow-crash-expected.txt

CONSOLE MESSAGE: line 23: PASS unless crash
